Fix write_data() when size 0
When working with RTP packets, if you receive a new network packet without payload then the relay loop is exited. This is not a good behaviour, as you can receive such packets at any time. Thus this patch fixes this permiting the write of 0 bytes inside the function. No side effects are detected, so the patch looks good.